test/web-platform-tests/to-upstream/xhr/access-control-preflight-request-header-returns-origin.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Access-Control-Request-Origin accept different origin between preflight and actual request</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script>
"use strict";

async_test(t => {
  const xhr = new XMLHttpRequest();

  xhr.open("GET", corsURL("resources/access-control-preflight-request-header-returns-origin.py"));

  xhr.setRequestHeader("X-Test", "foobar");

  xhr.onerror = t.unreached_func("Error occurred.");

  xhr.onload = t.step_func_done(() => {
    assert_equals(xhr.status, 200);
    assert_equals(xhr.responseText, "PASS");
  });

  xhr.send();
});

function corsURL(path) {
  const url = new URL(path, location.href);
  url.hostname = "www1." + url.hostname;
  return url.href;
}
</script>
